# refactor(Snowflake): simplify `compare` logic