This simplifies and optimizes same-type checks, which are likelier than mismatching types, the benchmarks below are run on 100,000,000 iterations:

| Pair               | Old      | New      | Comparison |
| :----------------: | :------: | :------: | :--------: |
| `string`, `string` | 43.623ms | 37.412ms |     85.76% |
| `string`, `bigint` |   8.242s |   8.576s |    104.05% |
| `bigint`, `bigint` |   2.028s |   1.904s |     93.88% |